to support mounted apis, we need to move the core rpc construction out, but since there is a lot that depends on muxrpc already, we want this to be backwards compatible, so we do not have to rewrite everything immediately.
